move ActiveSelectionStats to a new component
